Add unit test for colorize_url
